Refactor: Replace references to Infrastructure.Osm.Models with Application.Ports.Model in mapping and reader classes
