+++ src/POIneer.Render/Application/Mapping/RawPoiMapper.cs
@@ -1,5 +1,5 @@
 using POIneer.Render.Domain.Models;
-using POIneer.Render.Infrastructure.Osm.Models;
+using POIneer.Render.Application.Ports.Model;
 
 namespace POIneer.Render.Application.Mapping;
