+++ tests/POIneer.Render.UnitTests/Application/UseCases/RenderRegions/RenderRegionTests.cs
@@ -2,9 +2,9 @@ using Microsoft.Extensions.Logging;
 using NSubstitute;
 using POIneer.Render.Application.Contracts;
 using POIneer.Render.Application.Mapping;
+using POIneer.Render.Application.Ports.Model;
 using POIneer.Render.Application.UseCases;
 using POIneer.Render.Domain.Models;
-using POIneer.Render.Infrastructure.Osm.Models;
 using POIneer.Render.Ports;
 using POIneer.Render.TestHelpers;
 using Xunit;
